+++ CHANGELOG.md
@@ -34,7 +34,8 @@
 ##### Minecraft 1.19.4, 1.20[.X], 1.21[.X]
 #### Fix
 - It wasn't possible to join the server using an online account if `forced-offline-uuid` was enabled
-- Fix `hide-player-coords` not working for players with `ONLINE` status on `online-mode` servers [#251](https://github.com/NikitaCartes/EasyAuth/issues/251)
+- `hide-player-coords` not working for players with `ONLINE` status on `online-mode` servers [#251](https://github.com/NikitaCartes/EasyAuth/issues/251)
+- Packets are being handled for not authenticated players (#230)(https://github.com/NikitaCartes/EasyAuth/issues/230)
 
 ---
 ### 3.4.1